Article Components' Name Changed
