+++ client/src/assets/placeholder/widget.js
@@ -1,0 +1,22 @@
+export const ARTICLECARD = Object.freeze({
+  tags: ['Department', 'Campus', 'BM-BT'],
+  title: 'Chaos, Curiosity and COVID-19: A Biotechnologist’s Perspective',
+  authors: ['Debabrata Malik', 'Rama Krushna Behera'],
+  readTime: '12 min',
+  summary:
+    'A webinar on COVID-19 was conducted by the BM/BT Department. Read on to know more. A webinar on COVID-19 was conducted by the BM/BT Department. Read on to know more.',
+});
+
+export const POLLS = Object.freeze({
+  question: 'What is your opinion about the right time for return of students to the campus?',
+  votes: [
+    'I am desperately waiting to return back to campus with rigid safety protocols.',
+    'I am eager to return but I am a bit apprehensive about safety protocols.',
+    "I don't want to return in the near future because of the risk of transmission and would rather wait for normalcy.",
+  ],
+});
+
+export const SQUIGGLES = Object.freeze({
+  text:
+    'Hon’ble Union Education Minister inaugurated the Golden Jubilee Building at NIT Rourkela on 19th October. The broadcast of the inauguration ceremony can be watched with the following link: https://youtu.be/CTi4VnwdGVI',
+});
