+++ client/src/components/widgets/Pulse.js
@@ -4,18 +4,14 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { Card, CardContent, Grid, Typography, Button } from '@material-ui/core';
 
+// Components
+import { POLLS } from '../../assets/placeholder/widget';
+
 // images
 import pulseImg from '../../assets/images/pulseImg.png';
 
 const Pulse = () => {
-  const polls = {
-    question: 'What is your opinion about the right time for return of students to the campus?',
-    votes: [
-      'I am desperately waiting to return back to campus with rigid safety protocols.',
-      'I am eager to return but I am a bit apprehensive about safety protocols.',
-      "I don't want to return in the near future because of the risk of transmission and would rather wait for normalcy.",
-    ],
-  };
+  const polls = POLLS;
   const classes = useStyles();
   return (
     <Card className={classes.pulseCard}>
